+++ src/ScriptInterface/PlayingItemLocation.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "PlayingItemLocation.h"
 
-PlayingItemLocation::PlayingItemLocation(bool isValid, size_t playlistIndex, size_t playlistItemIndex) : m_isValid(isValid), m_playlistIndex(playlistIndex), m_playlistItemIndex(playlistItemIndex) {}
+PlayingItemLocation::PlayingItemLocation(bool isValid, uint32_t playlistIndex, uint32_t playlistItemIndex) : m_isValid(isValid), m_playlistIndex(playlistIndex), m_playlistItemIndex(playlistItemIndex) {}
 
 STDMETHODIMP PlayingItemLocation::get_IsValid(VARIANT_BOOL* out)
 {
